pyproject.toml
[project]
name = "small-cap-picker"
version = "0.1.0"
description = "Multiagent system for picking up the best small caps to invest."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.96.0",
    "finvizfinance>=1.3.0",
    "httpx>=0.28.1",
    "json-repair>=0.59.4",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "yfinance>=1.3.0",
]
